Map loads from a csv file.
Performs no bounds checking at this time.  Assumes that the file has
been formed correctly.

+++ Object.cpp
@@ -6,7 +6,7 @@ ResourceManager<sf::Texture> GLOBAL::textureList;
 //*************************************************************
 // Creates the sprite for the object.  
 // Sets the inital scale, source position, and screen position.
-void Object::loadObject(std::string filename, sf::Vector2f pos, sf::Vector2f sSize, sf::Vector2f sPos) {
+void Object::loadObject(std::string filename, sf::Vector2f pos, sf::Vector2i sSize, sf::Vector2i sPos) {
 	this->srcSize = srcSize;
 	this->srcPos = srcPos;
 	sprite = new sf::Sprite;
