+++ Player.h
@@ -15,7 +15,7 @@ class Player : public Object  {
 	std::vector<Card> hand;
 	Robot robot;
 public:
-	Player(std::string filename, sf::Vector2f pos, sf::Vector2f sPos, sf::Vector2f sSize = CARD_SOURCE_SIZE) 
+	Player(std::string filename, sf::Vector2f pos, sf::Vector2i sPos, sf::Vector2i sSize = CARD_SOURCE_SIZE) 
 		: Object(filename, pos, sPos, sSize), robot(filename, pos, sPos) {};
 
 	void programRobot() {};
