+++ Tile.h
@@ -15,7 +15,7 @@ class Tile : public Object {
 
 public:
 	Tile(std::string tileInformation, std::string spriteSheetFilename, sf::Vector2f pos /*, sf::Vector2f sPos, sf::Vector2f sSize = TILE_SOURCE_SIZE*/)    //commented out because i will find sPos through parsing etc
-		: Object(spriteSheetFilename, pos, sf::Vector2f(0, 0), TILE_SOURCE_SIZE) {};   //sf::Vector2f(0,0) is passing a (0,0) to the object constructor, needs to be replaced with sPos (source position)
+		: Object(spriteSheetFilename, pos, sf::Vector2i(0, 0), TILE_SOURCE_SIZE) {};   //sf::Vector2f(0,0) is passing a (0,0) to the object constructor, needs to be replaced with sPos (source position)
 
 };
