+++ TileFeature.h
@@ -8,7 +8,7 @@
 #include <string>
 #include "Object.h"
 
-const sf::Vector2f TILE_SOURCE_SIZE = { 150, 150 };
+const sf::Vector2i TILE_SOURCE_SIZE = { 150, 150 };
 
 class TileFeature : public Object  {
 	std::string featureName;
@@ -19,7 +19,7 @@ class TileFeature : public Object  {
 	bool moveRobot;
 
 public:
-	TileFeature(std::string filename, sf::Vector2f pos, sf::Vector2f sPos, sf::Vector2f sSize = TILE_SOURCE_SIZE) 
+	TileFeature(std::string filename, sf::Vector2f pos, sf::Vector2i sPos, sf::Vector2i sSize = TILE_SOURCE_SIZE) 
 		: Object(filename, pos, sPos, sSize) {};
 
 };
